+++ packages/analytics-js-service-worker/src/Analytics.ts
@@ -10,6 +10,7 @@ import ms from 'ms';
 import { v4 as uuid } from '@lukeed/uuid';
 import { is, clone } from 'ramda';
 import fetchAdapter from '@vespaiach/axios-fetch-adapter';
+import { isValidURL } from '@rudderstack/analytics-js-common/utilities/url';
 import type { IAnalytics } from './IAnalytics';
 import type {
   ApiCallback,
@@ -19,7 +20,7 @@ import type {
   IntegrationOptions,
 } from './types';
 import { looselyValidateEvent } from './loosely-validate-event';
-import { getDataPlaneUrl, isFunction, isValidUrl, noop } from './utilities';
+import { getDataPlaneUrl, isFunction, noop } from './utilities';
 
 const version = '__PACKAGE_VERSION__';
 
@@ -63,7 +64,7 @@ class Analytics implements IAnalytics {
       throw new Error('You must pass the source write key.');
     }
 
-    if (!isValidUrl(dataPlaneURL)) {
+    if (!isValidURL(dataPlaneURL)) {
       throw new Error(`The provided data plane URL "${dataPlaneURL}" is invalid.`);
     }